revert(terminal): remove unsupported React #185 parking fix
