+++ src/renderer/src/components/terminal-pane/TerminalPaneOverlayLayer.tsx
@@ -32,8 +32,7 @@ const TerminalPaneOverlayLayer = memo(function TerminalPaneOverlayLayer({
   shouldMeasureHiddenWorktree = false,
   activityTerminalPortals = EMPTY_ACTIVITY_PORTALS,
   backgroundMountTabIds = null,
-  activationDeferredMountTabIds = null,
-  onActivationDeferredWatcherHandoffFailed
+  activationDeferredMountTabIds = null
 }: {
   worktreeId: string
   worktreePath: string
@@ -47,7 +46,6 @@ const TerminalPaneOverlayLayer = memo(function TerminalPaneOverlayLayer({
   backgroundMountTabIds?: ReadonlySet<string> | null
   /** Cold-activation deferred tabs receive immediate parked watcher coverage. */
   activationDeferredMountTabIds?: ReadonlySet<string> | null
-  onActivationDeferredWatcherHandoffFailed?: (tabId: string) => void
 }): React.JSX.Element | null {
   const { terminalTabs, unifiedTabs, groups, activeGroupId } = useAppStore(
     useShallow((state) => ({
@@ -125,8 +123,7 @@ const TerminalPaneOverlayLayer = memo(function TerminalPaneOverlayLayer({
     isForceParked,
     shouldMeasureHiddenWorktree,
     activityTerminalPortals,
-    activationDeferredMountTabIds,
-    onActivationDeferredWatcherHandoffFailed
+    activationDeferredMountTabIds
   })
 
   if (!worktreePath) {
